refactor(1d): move GetNextChallengeForPix1d to school bounded context
