+++ api/src/school/domain/models/Activity.js
