+++ api/src/school/domain/services/algorithm-method.js
@@ -1,5 +1,5 @@
-import { Activity } from '../../models/Activity.js';
-import { logger } from '../../../infrastructure/logger.js';
+import { Activity } from '../models/Activity.js';
+import { logger } from '../../../../lib/infrastructure/logger.js';
 
 export const pix1dService = { getNextActivityLevel };
