+++ api/tests/integration/infrastructure/repositories/activity-repository_test.js
@@ -1,6 +1,6 @@
 import { catchErr, databaseBuilder, expect, knex } from '../../../test-helper.js';
 import * as activityRepository from '../../../../lib/infrastructure/repositories/activity-repository.js';
-import { Activity } from '../../../../lib/domain/models/Activity.js';
+import { Activity } from '../../../../src/school/domain/models/Activity.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 
 describe('Integration | Repository | activityRepository', function () {
