+++ api/tests/integration/infrastructure/repositories/challenge-repository_test.js
@@ -4,7 +4,7 @@ import { Challenge } from '../../../../lib/domain/models/Challenge.js';
 import { Validator } from '../../../../lib/domain/models/Validator.js';
 import * as challengeRepository from '../../../../lib/infrastructure/repositories/challenge-repository.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { Activity } from '../../../../lib/domain/models/Activity.js';
+import { Activity } from '../../../../src/school/domain/models/Activity.js';
 import * as learningContentBuilder from '../../../tooling/learning-content-builder/index.js';
 
 describe('Integration | Repository | challenge-repository', function () {
