+++ api/tests/unit/infrastructure/serializers/jsonapi/activity-serializer_test.js
@@ -1,6 +1,6 @@
 import { domainBuilder, expect } from '../../../../test-helper.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/activity-serializer.js';
-import { Activity } from '../../../../../lib/domain/models/Activity.js';
+import { Activity } from '../../../../../src/school/domain/models/Activity.js';
 
 describe('Unit | Serializer | JSONAPI | activity-serializer', function () {
   describe('#serialize', function () {
